Return float one-hot from LeafCTMC sampling
Use frequencies.dtype for the distribution dtype and the sampled
one-hot output, so dist.log_prob(dist.sample()) works without casting.
Previously the int32 samples were incompatible with the float
TensorArray in phylogenetic_likelihood.

+++ treeflow/distributions/leaf_ctmc.py
@@ -26,7 +26,7 @@ class LeafCTMC(Distribution):
             validate_args=validate_args,
             allow_nan_stats=allow_nan_stats,
             name=name,
-            dtype=tf.int32,
+            dtype=frequencies.dtype,
             reparameterization_type=NOT_REPARAMETERIZED,
             parameters=parameters,
         )
